Fix embed issue
If you're seeing this commit, I implore you to only grade my commit that was submitted before midnight on 11/14/24. This commit is simply meant to fix an issue with the embedded tableau chart, and I would rather take the deduction for not having the site work properly than the deduction for a late submission.